#12 refactor: user 패키지 변경 사항 반영
