+++ core/src/main/kotlin/org/doorip/core/auth/AccessTokenUseCase.kt
@@ -1,6 +1,6 @@
 package org.doorip.core.auth
 
-import org.doorip.domain.UserId
+import org.doorip.domain.entity.UserId
 
 interface AccessTokenUseCase {
     fun getUserId(token: String): UserId?
