--- domain/src/main/kotlin/org/doorip/domain/User.kt
@@ -1,8 +1,0 @@
-package org.doorip.domain
-
-@JvmInline
-value class UserId(
-    val value: Long,
-)
-
-class User
