+++ domain/src/main/kotlin/org/doorip/domain/entity/User.kt
@@ -1,0 +1,13 @@
+package org.doorip.domain.entity
+
+@JvmInline
+value class UserId(
+    val value: Long,
+)
+
+data class User(
+    val id: UserId,
+    val name: String,
+    val intro: String,
+    val result: Int?,
+)
