+++ presentation/support/auth/src/main/kotlin/org/doorip/presentation/support/auth/AuthenticatedUser.kt
@@ -1,6 +1,6 @@
 package org.doorip.presentation.support.auth
 
-import org.doorip.domain.UserId
+import org.doorip.domain.entity.UserId
 
 data class AuthenticatedUser(
     val userId: UserId,
