+++ presentation/support/auth/src/main/kotlin/org/doorip/presentation/support/auth/TokenAuthenticationFilter.kt
@@ -4,7 +4,7 @@ import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.doorip.core.auth.AccessTokenUseCase
-import org.doorip.domain.UserId
+import org.doorip.domain.entity.UserId
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource
 import org.springframework.stereotype.Component
